reorganise base transformers and functions, clean imports (#133)
