+++ feature_engine/discretisation/equal_frequency.py
@@ -2,10 +2,6 @@
 # License: BSD 3 clause
 
 import pandas as pd
-
-from sklearn.tree import DecisionTreeRegressor, DecisionTreeClassifier
-from sklearn.model_selection import GridSearchCV
-
 from feature_engine.variable_manipulation import _define_variables
 from feature_engine.base_transformers import BaseNumericalTransformer
