+++ feature_engine/encoding/woe.py
@@ -3,23 +3,10 @@
 
 import numpy as np
 import pandas as pd
-import warnings
 
-from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.pipeline import Pipeline
-from sklearn.utils.validation import check_is_fitted
-
-from feature_engine.dataframe_checks import _is_dataframe, _check_input_matches_training_df, _check_contains_na
+from feature_engine.dataframe_checks import _is_dataframe, _check_contains_na
 from feature_engine.variable_manipulation import _find_categorical_variables, _define_variables
-from feature_engine.base_transformers import BaseCategoricalTransformer
-
-
-def _check_encoding_dictionary(dictionary):
-    # check that there is a dictionary with category to number pairs
-    if len(dictionary) == 0:
-        raise ValueError('Encoder could not be fitted. Check the parameters and the variables '
-                         'in your dataframe.')
-    return dictionary
+from feature_engine.encoding.base_encoder import BaseCategoricalTransformer, _check_encoding_dictionary
 
 
 class WoEEncoder(BaseCategoricalTransformer):
@@ -104,8 +91,7 @@ class WoEEncoder(BaseCategoricalTransformer):
         # check input dataframe
         X = _is_dataframe(X)
 
-        # find categorical variables or check that those entered by the user
-        # are of type object
+        # find categorical variables or check that variables entered by user are of type object
         self.variables = _find_categorical_variables(X, self.variables)
 
         # check if dataset contains na
