+++ feature_engine/imputation/random_sample.py
@@ -7,13 +7,7 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
 from feature_engine.dataframe_checks import _is_dataframe, _check_input_matches_training_df
-from feature_engine.variable_manipulation import (
-    _find_categorical_variables,
-    _define_variables,
-    _find_numerical_variables,
-    _define_numerical_dict
-)
-from feature_engine.base_transformers import BaseImputer
+from feature_engine.variable_manipulation import _define_variables
 
 
 # for RandomSampleImputer
@@ -183,8 +177,7 @@ class RandomSampleImputer(BaseEstimator, TransformerMixin):
         # check that input is a dataframe
         X = _is_dataframe(X)
 
-        # Check that the dataframe contains the same number of columns than the dataframe
-        # used to fit the imputer.
+        # Check that input data contains same number of columns as dataframe used to fit
         _check_input_matches_training_df(X, self.input_shape_[1])
 
         # random sampling with a general seed
