+++ feature_engine/outliers/artbitrary.py
@@ -5,10 +5,7 @@ import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine.variable_manipulation import (
-    _define_variables,
-    _find_numerical_variables
-)
+from feature_engine.variable_manipulation import _find_numerical_variables
 
 from feature_engine.dataframe_checks import (
     _is_dataframe,
