+++ feature_engine/transformation/log.py
@@ -2,7 +2,6 @@
 # License: BSD 3 clause
 
 import numpy as np
-import scipy.stats as stats
 
 from feature_engine.variable_manipulation import _define_variables
 from feature_engine.base_transformers import BaseNumericalTransformer
